Refact: StateMachine add isTrigger to transitions
Now there is two types of transitions:

Real:
Which alter the stateMachine through stateTo or a function.

Trigger:
Which do not alter the stateMachine.

It only run functions that will afect an external
proccess. This proccess may will later send a message
to the stateMachine and then the stateMachine will
change it's state through a real transition.